Move the shared slug resolver where shared code lives
`suppressions.ts` was importing it from `supportRequests.ts`, so one
mutation module depended on another. Both are slug concerns and
`_profileSlugs.ts` already owns `readProfileSlugFromInput` and
`validateProfileSlug`, which this composes.

Also cap the message before normalizing rather than after. The mutation
is public and unauthenticated, so its argument size is whatever a caller
sends, and four regex passes ahead of the slice made that work scale with
what an attacker picks instead of with what is kept.
